Fix rendering far distances
Reverting part of commit 4e152ff4b8d7a56d8ff3f43197be2801a835856a. I noticed GivenRC1.wad (https://www.doomworld.com/forum/topic/95543-the-given-a-large-monsterless-puzzle-map-rc1/) wasn't rendering correctly, but in DR v2.5.3 it does.

+++ src/r_main.c
@@ -152,13 +152,13 @@ int R_PointOnSegSide(fixed_t x, fixed_t y, seg_t *line)
 
 static int SlopeDiv(unsigned int num, unsigned int den)
 {
-    unsigned int ans;
+    uint64_t    ans;
 
     if (den < 512)
         return (ANG45 - 1);
 
-    ans = (num << 3) / (den >> 8);
-    return (ans <= SLOPERANGE ? tantoangle[ans] : (ANG45 - 1));
+    ans = ((uint64_t)num << 3) / (den >> 8);
+    return (int)(ans <= SLOPERANGE ? tantoangle[ans] : (ANG45 - 1));
 }
 
 //
